RavenDB-24991 Move resources to chatbot
