--- src/Raven.Studio/typescript/components/common/helpAndResources/HelpAndResourcesWidget.tsx
@@ -1,96 +1,0 @@
-import React, { useState } from "react";
-import { Icon } from "components/common/Icon";
-import "./HelpAndResources.scss";
-import { FlexGrow } from "components/common/FlexGrow";
-import { JoinCommunityContent } from "components/common/helpAndResources/partials/JoinCommunityContent";
-import { FeedbackFormContent } from "components/common/helpAndResources/partials/FeedbackFormContent";
-import { ContactSupportContent } from "components/common/helpAndResources/partials/ContactSupportContent";
-import useBoolean from "components/hooks/useBoolean";
-import GithubButtonWithStars from "components/common/helpAndResources/partials/GithubButtonWithStars";
-import SeeDocumentationButton from "components/common/helpAndResources/partials/SeeDocumentationButton";
-import PopoverWithHoverWrapper from "components/common/PopoverWithHoverWrapper";
-
-type Tab = "Help and resources" | "Join the community" | "Contact support" | "Submit feedback";
-
-export function HelpAndResourcesWidget() {
-    const { value: isMenuOpen, setValue: setIsMenuOpen } = useBoolean(false);
-    const [activeTab, setActiveTab] = useState<Tab>("Help and resources");
-
-    const toggleIsMenuOpen = () => {
-        setIsMenuOpen(!isMenuOpen);
-        if (isMenuOpen) {
-            setActiveTab("Help and resources");
-        }
-    };
-
-    return (
-        <section className="bottom-right-tools-container">
-            {!isMenuOpen && (
-                <PopoverWithHoverWrapper message="Help and resources">
-                    <span role="button" className="btn-help-widget" onClick={toggleIsMenuOpen}>
-                        <Icon icon="question" margin="m-0" />
-                    </span>
-                </PopoverWithHoverWrapper>
-            )}
-            {isMenuOpen && (
-                <div className="action-menu">
-                    <span role="button" className="btn-hide-widget" onClick={toggleIsMenuOpen} title="Hide widget">
-                        <Icon icon="close" margin="m-0" className="lh-base" />
-                    </span>
-                    <div className="action-menu__header gap-2">
-                        {activeTab !== "Help and resources" && (
-                            <span
-                                role="button"
-                                className="action-menu__back"
-                                onClick={() => setActiveTab("Help and resources")}
-                                title="Back to Help and resources"
-                            >
-                                <Icon icon="arrow-thin-left" margin="m-0" />
-                            </span>
-                        )}
-                        <h4 className="lh-base">{activeTab}</h4>
-                    </div>
-                    {activeTab === "Help and resources" && (
-                        <>
-                            <ul className="action-menu__list">
-                                <li
-                                    className="action-menu__list-item action-menu__list-item--highlight"
-                                    role="button"
-                                    title="Join the Community"
-                                    onClick={() => setActiveTab("Join the community")}
-                                >
-                                    <Icon icon="group" margin="m-0" />
-                                    Join the Community
-                                </li>
-                                <li
-                                    className="action-menu__list-item"
-                                    role="button"
-                                    title="Contact support"
-                                    onClick={() => setActiveTab("Contact support")}
-                                >
-                                    <Icon icon="support" margin="m-0" />
-                                    Contact support
-                                </li>
-                                <SeeDocumentationButton />
-                            </ul>
-                            <div className="action-menu__footer">
-                                <div className="d-flex flex-row align-items-center">
-                                    <small className="lh-1">See our project on GitHub</small>
-                                    <FlexGrow />
-                                    <GithubButtonWithStars />
-                                </div>
-                            </div>
-                        </>
-                    )}
-                    {activeTab === "Join the community" && (
-                        <JoinCommunityContent openFeedbackForm={() => setActiveTab("Submit feedback")} />
-                    )}
-                    {activeTab === "Submit feedback" && (
-                        <FeedbackFormContent goBack={() => setActiveTab("Help and resources")} />
-                    )}
-                    {activeTab === "Contact support" && <ContactSupportContent />}
-                </div>
-            )}
-        </section>
-    );
-}
