+++ src/Raven.Studio/typescript/components/shell/chatbot/partials/ChatbotPanelNews.tsx
@@ -1,0 +1,3 @@
+export default function ChatbotPanelNews() {
+    return <div>News</div>;
+}
