+++ src/Raven.Studio/typescript/components/shell/chatbot/partials/ChatbotPanelWhatsNew.tsx
@@ -1,0 +1,3 @@
+export default function ChatbotPanelWhatsNew() {
+    return <div>Whats New</div>;
+}
