+++ src/Raven.Studio/typescript/components/shell/chatbot/partials/resources/partials/SeeDocumentationButton.tsx
