+++ src/Raven.Studio/wwwroot/App/views/shell.html
@@ -182,8 +182,6 @@
     <div id="page-host-root" class="content-container">
         <div class="d-flex h-100">
             <div id="page-host" class="page-host" data-bind="router: { cacheViews: false }"></div>
-            <!-- TODO -->
-            <!-- <div data-bind="react: $root.helpAndResourcesWidgetView, visible: $root.isHelpAndResourcesWidgetVisible" class="bs5" style="z-index: 9999;"></div> -->
             <div data-bind="react: $root.chatbotView" class="bs5"></div>
         </div>
     </div>
